requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow>=9.5.0
python-dotenv>=1.0.0
drf-yasg>=1.21.7
django-cors-headers>=4.3.0
django-filter>=23.5
requests>=2.31.0
python-dateutil>=2.8.2